+++ python/lsst/ts/authorize/handler/rest_authorize_handler.py
@@ -25,6 +25,7 @@ import asyncio
 import logging
 import os
 import types
+import typing
 from collections.abc import Iterable
 from http import HTTPStatus
 from types import TracebackType
@@ -39,11 +40,14 @@ from .base_authorize_handler import BaseAuthorizeHandler
 __all__ = [
     "RestAuthorizeHandler",
     "RestMessageType",
+    "AUTHENTICATION_ERROR_CODE",
     "AUTHLISTREQUEST_ENDPOINT",
     "GET_TOKEN_ENDPOINT",
     "ID_EXECUTE_PARAMS",
 ]
 
+AUTHENTICATION_ERROR_CODE = 1
+GENERAL_ERROR_CODE = 2
 AUTHLISTREQUEST_ENDPOINT = "/manager/api/authlistrequest/"
 AUTHORIZED_PENDING_PARAMS = (
     f"?status=Authorized&execution_status={ExecutionStatus.PENDING}"
@@ -90,10 +94,11 @@ class RestAuthorizeHandler(BaseAuthorizeHandler):
     def __init__(
         self,
         domain: salobj.Domain,
+        callback: typing.Callable[[int, str], typing.Awaitable[None]],
         log: logging.Logger | None = None,
         config: types.SimpleNamespace | None = None,
     ) -> None:
-        super().__init__(domain=domain, log=log, config=config)
+        super().__init__(domain=domain, log=log, config=config, callback=callback)
         assert self.config is not None
         self.response: None | RestMessageType | Iterable[RestMessageType] = None
         self.authlistrequest_url = (
@@ -113,6 +118,11 @@ class RestAuthorizeHandler(BaseAuthorizeHandler):
                 "Please set AUTHLIST_USER_NAME and AUTHLIST_USER_PASS environment variables."
             )
 
+    async def _handle_error(self, code: int, report: str) -> RestMessageType:
+        assert self.callback is not None
+        await self.callback(code=code, report=report)
+        return {"error": report}
+
     async def _get_response(
         self, resp: aiohttp.ClientResponse
     ) -> RestMessageType | Iterable[RestMessageType]:
@@ -120,10 +130,12 @@ class RestAuthorizeHandler(BaseAuthorizeHandler):
             return await resp.json()
         else:
             resp_json = await resp.json()
-            raise RuntimeError(
+            code = GENERAL_ERROR_CODE
+            report = (
                 f"Got HTTP response status {resp.status} == {HTTPStatus(resp.status).name} "
                 f"and {resp_json=!s}."
             )
+            return await self._handle_error(code=code, report=report)
 
     async def authenticate(self) -> None:
         """Authenticate against the REST server.
@@ -148,7 +160,8 @@ class RestAuthorizeHandler(BaseAuthorizeHandler):
             else:
                 self.token = ""
                 self.log.error("Authentication unsuccessful.")
-                raise RuntimeError(f"Got unexpected response {self.response}.")
+                report = f"Got unexpected response {self.response}."
+                await self._handle_error(code=AUTHENTICATION_ERROR_CODE, report=report)
 
     async def handle_authorize_request(self, data: AuthRequestData) -> None:
         """Send a POST with the authorize request data.
@@ -172,7 +185,7 @@ class RestAuthorizeHandler(BaseAuthorizeHandler):
         async with self.lock, self.client_session.post(
             url=self.authlistrequest_url,
             json=json,
-            headers={"Authorization": self.token},
+            headers={"Authorization": f"Token {self.token}"},
         ) as resp:
             self.response = await self._get_response(resp=resp)
 
@@ -187,7 +200,7 @@ class RestAuthorizeHandler(BaseAuthorizeHandler):
         await self.authenticate()
         async with self.lock, self.client_session.get(
             self.authlistrequest_url + AUTHORIZED_PENDING_PARAMS,
-            headers={"Authorization": self.token},
+            headers={"Authorization": f"Token {self.token}"},
         ) as resp:
             self.response = await self._get_response(resp=resp)
             if self.response is not None:
@@ -226,7 +239,7 @@ class RestAuthorizeHandler(BaseAuthorizeHandler):
                             "execution_status": execution_status,
                             "execution_message": execution_message,
                         },
-                        headers={"Authorization": self.token},
+                        headers={"Authorization": f"Token {self.token}"},
                     ) as put_resp:
                         put_resp_json = await self._get_response(resp=put_resp)
                         assert isinstance(put_resp_json, dict)  # keep MyPy happy.